Added the math dataset test file with about 100 Questions and Answers
This file contains 100 Questions and Answers regarding popular Math topics based on the South African CAPS (Curriculum and Assessment Policy Statement) to ensure a large variety of Math Assistance aimed to South African students solely . This is the file I created to test the knowledge base that the SAiL Chatbot will be starting with. To test the dataset I used Google Colabs free cloud platform to run Python Scripts to test the questions and ensure the AI Model outputs the correct expected response and to ensure accurate answering and interaction.